Add files via upload
The developed "Diabetic Retinopathy using Neural Network" (DRuNN) model helps to detects the OCT based scan is DR affected patients. If a person is suffering from blurred vision, floating spot in eyes and diabetes, the physician will advise the patient to take the OCT scan and if the OCT scan consists the red color dots and yellowish-white color physician will take the decision that the OCT scan is of DR affected patient. So in this proposed research work, DRuNN technique is developed which detects whether a person is suffering from DR or not. With the help of CNN and ReLU, the model helps to understand the curve and color based feature of the scan. The achieved accuracy, error, specificity, and sensitivity of DRuNN techniques are 90.00%, 10.00%, 86.67%, and 94.28% respectively when 80% data is reserved for training phase and rest 20% data is reserved for testing phase.